SeqBox HT16K33 Fix
LED表示,Switch読み取りタイミング調整済

+++ TBSEQ_Board_V1/SeqBox/src/mode/matrixLED.cpp
@@ -7,8 +7,8 @@ matrixLED::matrixLED(){
         }
 }
 
-void matrixLED::setRow(int row, uint8_t value) {
-    uint8_t _maskbit;
+void matrixLED::setRow(int row, uint16_t value) {
+    uint16_t _maskbit;
     int _rowIndex = row * LED_COL_MAX;
     if (_rowIndex >= LED_INDEX_MAX) {
         _rowIndex = 0;
@@ -20,8 +20,8 @@ void matrixLED::setRow(int row, uint8_t value) {
     }
 }
 
-uint8_t matrixLED::getRow(int row) {
-    uint8_t _ret = 0x00;
+uint16_t matrixLED::getRow(int row) {
+    uint16_t _ret = 0x00;
     int _rowIndex = row * LED_COL_MAX;
 
     if (_rowIndex >= LED_INDEX_MAX) {
@@ -30,12 +30,12 @@ uint8_t matrixLED::getRow(int row) {
 
     //bit16-1を_retに設定
     for (int i = LED_COL_MAX - 1;1 <= i;i--) {
-        _ret = _ret + (uint8_t)_currentLED[_rowIndex + i];
+        _ret = _ret + (uint16_t)_currentLED[_rowIndex + i];
         _ret = _ret << 1;
     }
 
     //bit0を_retに設定
-    _ret = _ret + (uint8_t)_currentLED[_rowIndex];
+    _ret = _ret + (uint16_t)_currentLED[_rowIndex];
 
     return _ret;
 }
